Moved get_input into shared
