+++ src/os/macos.rs
@@ -1,29 +1,9 @@
-use ::pnet::datalink::Channel::Ethernet;
-use ::pnet::datalink::DataLinkReceiver;
-use ::pnet::datalink::{self, Config, NetworkInterface};
-use ::std::io::{self, stdin, Write};
-use ::termion::event::Event;
-use ::termion::input::TermRead;
-
 use ::std::collections::HashMap;
-use ::std::net::IpAddr;
-
-use signal_hook::iterator::Signals;
 
 use crate::network::Connection;
-use crate::OsInputOutput;
 
 use super::lsof_utils;
 use std::net::SocketAddr;
-
-use crate::os::shared::{
-    KeyboardEvents,
-    get_datalink_channel,
-    get_interface,
-    lookup_addr,
-    sigwinch,
-    create_write_to_stdout,
-};
 
 #[derive(Debug)]
 struct RawConnection {
@@ -34,7 +14,7 @@ struct RawConnection {
     process_name: String,
 }
 
-fn get_open_sockets() -> HashMap<Connection, String> {
+pub(crate) fn get_open_sockets() -> HashMap<Connection, String> {
     let mut open_sockets = HashMap::new();
 
     let connections = lsof_utils::get_connections();
@@ -52,29 +32,4 @@ fn get_open_sockets() -> HashMap<Connection, String> {
     }
 
     return open_sockets;
-}
-
-pub fn get_input(interface_name: &str) -> Result<OsInputOutput, failure::Error> {
-    let keyboard_events = Box::new(KeyboardEvents);
-    let network_interface = match get_interface(interface_name) {
-        Some(interface) => interface,
-        None => {
-            failure::bail!("Cannot find interface {}", interface_name);
-        }
-    };
-    let network_frames = get_datalink_channel(&network_interface)?;
-    let lookup_addr = Box::new(lookup_addr);
-    let write_to_stdout = create_write_to_stdout();
-    let (on_winch, cleanup) = sigwinch();
-
-    Ok(OsInputOutput {
-        network_interface,
-        network_frames,
-        get_open_sockets,
-        keyboard_events,
-        lookup_addr,
-        on_winch,
-        cleanup,
-        write_to_stdout,
-    })
 }
